# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/issues": {
      "get": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "search_issues",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "project_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "jql",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "create_issue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Issue created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{id}": {
      "get": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "get_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Issue id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "delete_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "update_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Issue id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{id}/restore": {
      "post": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "restore_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Issue id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/issues/{id}/trash": {
      "delete": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "purge_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Issue id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/projects/{key}/trash": {
      "get": {
        "tags": [
          "routes::issues"
        ],
        "operationId": "list_trash",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_key}/board": {
      "get": {
        "tags": [
          "routes::board"
        ],
        "operationId": "get_board",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardColumnResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "issue_ids"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "issue_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "wip_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "BoardResponse": {
        "type": "object",
        "required": [
          "project_id",
          "project_key",
          "columns",
          "issues",
          "sprint"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardColumnResponse"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            }
          },
          "project_id": {
            "type": "string"
          },
          "project_key": {
            "type": "string"
          },
          "sprint": {
            "$ref": "#/components/schemas/SprintResponse"
          }
        }
      },
...
      "CreateIssueRequest": {
        "type": "object",
        "required": [
          "project_key",
          "issue_type",
          "summary",
          "priority"
        ],
        "properties": {
          "assignee_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "issue_type": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "project_key": {
            "type": "string"
          },
          "reporter_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "IssueListResponse": {
        "type": "object",
        "required": [
          "issues"
        ],
        "properties": {
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            }
          }
        }
      },
      "IssueResponse": {
        "type": "object",
        "required": [
          "id",
          "key",
          "summary",
          "description",
          "issue_type",
          "project_key",
          "status",
          "status_id",
          "priority",
          "labels",
          "reporter_id",
          "project_name"
        ],
        "properties": {
          "assignee_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "assignee_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "issue_type": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "original_estimate_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Original estimate in seconds. Null when no estimate was set."
          },
          "priority": {
            "type": "string"
          },
          "project_key": {
            "type": "string"
          },
          "project_name": {
            "type": "string"
          },
          "remaining_estimate_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Remaining estimate in seconds, derived from the latest worklog."
          },
          "reporter_id": {
            "type": "string"
          },
          "reporter_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "sprint_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "status_id": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "SprintResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "goal",
          "state",
          "velocity",
          "issue_ids"
        ],
        "properties": {
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Format: date-time"
          },
          "goal": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "issue_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "remaining_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Format: date-time"
          },
          "state": {
            "type": "string"
          },
          "velocity": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdateIssueRequest": {
        "type": "object",
        "properties": {
          "affected_version_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "assignee_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "component_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fix_version_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "type": [
              "string",
              "null"
            ]
          },
          "sprint_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}